requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pgvector
pypdf
sentence-transformers
groq
python-multipart
aiofiles
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
greenlet
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
aiosqlite
psycopg2-binary
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2